# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AuditLog struct {
	ID          pgtype.UUID        `json:"id"`
	GroupID     pgtype.UUID        `json:"group_id"`
	ActorUserID pgtype.UUID        `json:"actor_user_id"`
	Action      string             `json:"action"`
	EntityType  string             `json:"entity_type"`
	EntityID    pgtype.UUID        `json:"entity_id"`
	OldData     []byte             `json:"old_data"`
	NewData     []byte             `json:"new_data"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
	DeletedAt   pgtype.Timestamptz `json:"deleted_at"`
}

type Customer struct {
	ID        pgtype.UUID        `json:"id"`
	GroupID   pgtype.UUID        `json:"group_id"`
	Phone     string             `json:"phone"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type EmployeeBalance struct {
	ID         pgtype.UUID        `json:"id"`
	GroupID    pgtype.UUID        `json:"group_id"`
	MemberID   pgtype.UUID        `json:"member_id"`
	BalanceUsd int64              `json:"balance_usd"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
	DeletedAt  pgtype.Timestamptz `json:"deleted_at"`
}

type Group struct {
	ID        pgtype.UUID        `json:"id"`
	Name      string             `json:"name"`
	CreatedBy pgtype.UUID        `json:"created_by"`
	IsActive  bool               `json:"is_active"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type GroupInvitation struct {
	ID            pgtype.UUID        `json:"id"`
	GroupID       pgtype.UUID        `json:"group_id"`
	InvitedBy     pgtype.UUID        `json:"invited_by"`
	InvitedUserID pgtype.UUID        `json:"invited_user_id"`
	LocationName  pgtype.Text        `json:"location_name"`
	Email         pgtype.Text        `json:"email"`
	Phone         pgtype.Text        `json:"phone"`
	Role          interface{}        `json:"role"`
	TokenHash     pgtype.Text        `json:"token_hash"`
	Status        string             `json:"status"`
	ExpiresAt     pgtype.Timestamptz `json:"expires_at"`
	AcceptedAt    pgtype.Timestamptz `json:"accepted_at"`
	RejectedAt    pgtype.Timestamptz `json:"rejected_at"`
	RespondedAt   pgtype.Timestamptz `json:"responded_at"`
	RevokedAt     pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
	DeletedAt     pgtype.Timestamptz `json:"deleted_at"`
}

type GroupMember struct {
	ID        pgtype.UUID        `json:"id"`
	GroupID   pgtype.UUID        `json:"group_id"`
	UserID    pgtype.UUID        `json:"user_id"`
	Username  string             `json:"username"`
	Role      interface{}        `json:"role"`
	IsOwner   bool               `json:"is_owner"`
	JoinedAt  pgtype.Timestamptz `json:"joined_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type Location struct {
	ID         pgtype.UUID        `json:"id"`
	GroupID    pgtype.UUID        `json:"group_id"`
	Name       string             `json:"name"`
	EmployeeID pgtype.UUID        `json:"employee_id"`
	CreatedBy  pgtype.UUID        `json:"created_by"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
	DeletedAt  pgtype.Timestamptz `json:"deleted_at"`
}

type MemberProfitPeriod struct {
	ID        pgtype.UUID        `json:"id"`
	GroupID   pgtype.UUID        `json:"group_id"`
	MemberID  pgtype.UUID        `json:"member_id"`
	Year      int16              `json:"year"`
	Month     int16              `json:"month"`
	ProfitUzs int64              `json:"profit_uzs"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type Notification struct {
	ID        pgtype.UUID        `json:"id"`
	TitleUz   string             `json:"title_uz"`
	TitleRu   string             `json:"title_ru"`
	TitleEn   string             `json:"title_en"`
	ContentUz string             `json:"content_uz"`
	ContentRu string             `json:"content_ru"`
	ContentEn string             `json:"content_en"`
	Type      interface{}        `json:"type"`
	Payload   []byte             `json:"payload"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type NotificationRecipient struct {
	ID             pgtype.UUID        `json:"id"`
	NotificationID pgtype.UUID        `json:"notification_id"`
	UserID         pgtype.UUID        `json:"user_id"`
	IsRead         bool               `json:"is_read"`
	ReadAt         pgtype.Timestamptz `json:"read_at"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
	DeletedAt      pgtype.Timestamptz `json:"deleted_at"`
}

type Order struct {
	ID        pgtype.UUID        `json:"id"`
	GroupID   pgtype.UUID        `json:"group_id"`
	Total     pgtype.Numeric     `json:"total"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Email        pgtype.Text        `json:"email"`
	Password     pgtype.Text        `json:"password"`
	PasswordHash pgtype.Text        `json:"password_hash"`
	Name         pgtype.Text        `json:"name"`
	Phone        pgtype.Text        `json:"phone"`
	Username     pgtype.Text        `json:"username"`
	FirstName    pgtype.Text        `json:"first_name"`
	LastName     pgtype.Text        `json:"last_name"`
	AvatarUrl    pgtype.Text        `json:"avatar_url"`
	Language     string             `json:"language"`
	IsActive     bool               `json:"is_active"`
	LastLoginAt  pgtype.Timestamptz `json:"last_login_at"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
	DeletedAt    pgtype.Timestamptz `json:"deleted_at"`
}